changes to request mapping end point

+++ src/main/java/com/careerdevs/gorestfinal3/repos/ToDoRepository.java
@@ -3,6 +3,6 @@ package com.careerdevs.gorestfinal3.repos;
 import com.careerdevs.gorestfinal3.models.ToDo;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ToDoRepository extends CrudRepository<ToDo, Integer> {
+public interface ToDoRepository extends CrudRepository<ToDo, Long> {
 
 }
